serial-core: keep value of console cflag in uart_update_termios
uart_update_termios is called on console resume and restores the line
parameters using cons->cflag before resetting it. However code in
uart_resume_port does the same afterwards, applying a 0 value to the
port's cflag and erasing the correct setting. This patch preserves the
value of cflag in uart_update_termios to prevent this and allow the
serial console to restore properly even if there is no program using it
on user side.

Serial code has been largely refactored upstream in the upstream kernel
and faulty code apparently removed.

Fixes bug 831077

Change-Id: Ia894f971d9b6bb0c6103dbca9d169dc2a2148f84
Reviewed-on: http://git-master/r/33098
Reviewed-by: Alex Courbot <acourbot@nvidia.com>
Tested-by: Alex Courbot <acourbot@nvidia.com>
Reviewed-by: David Schalig <dschalig@nvidia.com>
Reviewed-by: Janne Hellsten <jhellsten@nvidia.com>