tegra: usb: fix ehci suspend in host mode
When suspend is called and usb1 is in host mode, tegra_hci_remove is
invoked to remove the device. However, at this time it is likely that
runtime PM will have powered off usb1. This patch ensures usb1 is
powered on if necessary within tegra_hci_remove so that ehci controller
registers can be manipulated.

Bug 835677.

Change-Id: Ida9585ecb4fcc9a88786131a1b629905efd1a07a
Reviewed-on: http://git-master/r/35592
Reviewed-by: Rohan Somvanshi <rsomvanshi@nvidia.com>
Tested-by: Rohan Somvanshi <rsomvanshi@nvidia.com>